EC-235 - Updated ConflictController.update to first copy the conflict to the receiver_sync_archive table before marking it as resolved

+++ web-common/src/main/java/org/openmrs/eip/web/contoller/BaseRestController.java
@@ -47,7 +47,7 @@ public abstract class BaseRestController {
 		return results;
 	}
 	
-	public Object doGet(Integer id) {
+	public Object doGet(Long id) {
 		return producerTemplate.requestBody(
 		    "jpa:" + getName() + "?query=SELECT c FROM " + getName() + " c WHERE c.id = " + id, null, getClazz());
 	}
