+++ web-sender/src/main/java/org/openmrs/eip/web/sender/DebeziumEventController.java
@@ -29,7 +29,7 @@ public class DebeziumEventController extends BaseRestController {
 	}
 	
 	@GetMapping("/{id}")
-	public Object get(@PathVariable("id") Integer id) {
+	public Object get(@PathVariable("id") Long id) {
 		if (log.isDebugEnabled()) {
 			log.debug("Fetching debezium event with id: " + id);
 		}
