+++ web-sender/src/main/java/org/openmrs/eip/web/sender/SenderErrorController.java
@@ -33,7 +33,7 @@ public class SenderErrorController extends BaseRestController {
 	}
 	
 	@GetMapping("/{id}")
-	public Object get(@PathVariable("id") Integer id) {
+	public Object get(@PathVariable("id") Long id) {
 		if (log.isDebugEnabled()) {
 			log.debug("Fetching sender retry item with id: " + id);
 		}
